core-mem-fifo: pack structs
Modify and reorder structs so that the most frequently accessed attributes
are on the first cache line. Remove the useless start attribute from
mem_page_t. The only lost information is the total number of blocks in the
pool, which is hence not logged anymore when the pool is destroyed before
all the blocks are freed. Only the number of pages is displayed.

In mem_pool_t, the modification of the type of min_alignment allows to
reduce the size of the struct by 8 bytes. min_alignement is always less
than 16 so this shouldn't be a problem. In turn, this allows to keep all
the most used attributes of mem_fifo_pool_t (up to alive) in the first
cache line. This gives a 4-5% performance boost in the bench.

In mem_page_t, start was an alias to area and was never used.

+++ core-mem.h
@@ -239,7 +239,7 @@ enum mem_pools_t {
 
 typedef struct mem_pool_t {
     mem_flags_t mem_pool;
-    size_t      min_alignment;
+    uint32_t    min_alignment;
     struct mem_pool_t *realloc_fallback;
 
     /* DO NOT USE DIRECTLY, use mp_imalloc/mp_irealloc/mp_ifree instead */
